backend/pyproject.toml
[tool.poetry]
name = "challenge-constancia"
version = "0.1.0"
description = ""
authors = ["SMarkus27 <pereira.markus.27@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.29.0"
fastapi = "^0.110.1"
python-decouple = "^3.8"
psycopg = {extras = ["binary"], version = "^3.1.18"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
